requirements.txt
boto3==1.34.108
botocore==1.34.108
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cycler==0.11.0
dnspython==2.3.0
docopt==0.6.2
filelock==3.12.2
fonttools==4.38.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.5.3
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu116==11.6.124
nvidia-cuda-runtime-cu116==11.6.55
nvidia-cudnn-cu11==8.5.0.96
nvidia-pyindex==1.0.9
packaging==24.0
pandas==2.0.3
parallel-ssh==2.12.0
patsy==0.5.6
pillow==10.3.0
protobuf==5.26.1
psutil==5.9.8
pybind11==2.12.0
pyparsing==3.1.2
pysqlite3==0.5.2
python-dateutil==2.9.0.post0
python-etcd==0.4.5
pytz==2024.1
regex==2024.5.15
requests==2.31.0
s3transfer==0.10.1
scipy==1.10.1
six==1.16.0
ssh-python==1.0.0
ssh2-python==1.0.0
statsmodels==0.14.1
tensorboardX==1.8
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
yarg==0.1.9
zope.event==5.0
zope.interface==6.4.post2
